requirements.txt
pytest>=7.4.3
pytest-cov>=6.0.0
pytest-asyncio>=0.21.1
pyyaml>=6.0.1
litellm>=1.61.3
chromadb>=0.6.3
openai>=1.63.0
anthropic>=0.18.1
python-dotenv>=1.0.0
tiktoken>=0.9.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
